+++ client/src/DashboardPage.js
@@ -1,5 +1,5 @@
 import Vue from 'vue'
-import DashboardPage from './components/dashboard-page'
+import DashboardPage from './components/DashboardPage'
 
 Vue.config.productionTip = false;
