+++ client/src/components/TheInfo.vue
